Refactor: Moved NewGame and  HTP screens to own components
